requirements.txt
openai
python-dotenv
pyyaml
prompt_toolkit
tiktoken
requests
pandas
numpy
sympy
scikit-learn
matplotlib
beautifulsoup4
scipy
cloudscraper